requirements.txt
auto-gptq --extra-index-url https://huggingface.github.io/autogptq-index/whl/cu118/
bitsandbytes
transformers 
peft 
accelerate 
optimum
openai
autoawq
autoawq-kernels
